Move AppearancePrefsTab to gui package
